Use proper uint32 timestamp for bitcoin

+++ src/oracles/bitcoin/BitcoinOracle.sol
@@ -218,7 +218,7 @@ contract BitcoinOracle is BaseOracle {
      */
     function _getTimestampOfBlock(
         bytes calldata blockHeader
-    ) internal pure returns (uint256 timestamp) {
+    ) internal pure returns (uint32 timestamp) {
         return timestamp = Endian.reverse32(uint32(bytes4(blockHeader[68:68 + 4])));
     }
 
@@ -235,7 +235,7 @@ contract BitcoinOracle is BaseOracle {
     function _getTimestampOfPreviousBlock(
         bytes calldata previousBlockHeader,
         bytes calldata currentBlockHeader
-    ) internal pure returns (uint256 timestamp) {
+    ) internal pure returns (uint32 timestamp) {
         bytes32 proposedPreviousBlockHash = BtcProof.getBlockHash(previousBlockHeader);
         bytes32 actualPreviousBlockHash = bytes32(Endian.reverse256(uint256(bytes32(currentBlockHeader[4:36]))));
         if (actualPreviousBlockHash != proposedPreviousBlockHash) {
@@ -384,9 +384,9 @@ contract BitcoinOracle is BaseOracle {
         uint256 blockNum,
         BtcTxProof calldata inclusionProof,
         uint256 txOutIndex,
-        uint256 timestamp
+        uint32 timestamp
     ) internal {
-        if (timestamp + CAN_VALIDATE_OUTPUTS_FOR < block.timestamp) revert TooLate();
+        if (timestamp + CAN_VALIDATE_OUTPUTS_FOR < uint32(block.timestamp)) revert TooLate();
 
         bytes32 token = output.token;
         uint256 sats = _validateUnderlyingPayment(
@@ -399,14 +399,14 @@ contract BitcoinOracle is BaseOracle {
         );
         if (sats != output.amount) revert BadAmount(); // Exact amount is checked to protect against "double spends".
 
-        bytes32 solver = _resolveClaimed(uint32(timestamp), orderId, output);
+        bytes32 solver = _resolveClaimed(timestamp, orderId, output);
 
         bytes32 outputHash =
-            keccak256(MandateOutputEncodingLib.encodeFillDescription(solver, orderId, uint32(timestamp), output));
+            keccak256(MandateOutputEncodingLib.encodeFillDescription(solver, orderId, timestamp, output));
         _attestations[block.chainid][bytes32(uint256(uint160(address(this))))][bytes32(uint256(uint160(address(this))))][outputHash]
         = true;
 
-        emit OutputFilled(orderId, solver, uint32(timestamp), output);
+        emit OutputFilled(orderId, solver, timestamp, output);
         emit OutputVerified(inclusionProof.txId);
     }
 
@@ -426,7 +426,7 @@ contract BitcoinOracle is BaseOracle {
         uint256 txOutIndex
     ) internal {
         // _validateUnderlyingPayment checks if inclusionProof.blockHeader == 80.
-        uint256 timestamp = _getTimestampOfBlock(inclusionProof.blockHeader);
+        uint32 timestamp = _getTimestampOfBlock(inclusionProof.blockHeader);
         _verify(orderId, output, blockNum, inclusionProof, txOutIndex, timestamp);
     }
 
@@ -448,7 +448,7 @@ contract BitcoinOracle is BaseOracle {
         bytes calldata previousBlockHeader
     ) internal {
         // _validateUnderlyingPayment checks if inclusionProof.blockHeader == 80.
-        uint256 timestamp = _getTimestampOfPreviousBlock(previousBlockHeader, inclusionProof.blockHeader);
+        uint32 timestamp = _getTimestampOfPreviousBlock(previousBlockHeader, inclusionProof.blockHeader);
         _verify(orderId, output, blockNum, inclusionProof, txOutIndex, timestamp);
     }
